refactor: 쿠폰관리 위젯 분리
