+++ src/refactoring/widget/coupon-management/index.ts
@@ -1,0 +1,1 @@
+export { CouponManagement } from "./ui";
